requirements.txt
pandas
numpy
pyyaml
pyarrow
